+++ src/components/internal/AnchorButton/AnchorButton.tsx
@@ -1,6 +1,6 @@
 import { Ref } from 'preact';
 import useButton from '../Button/hooks/useButton';
-import { DEFAULT_BUTTON_CLASSNAME } from '../Button/constants';
+import { DEFAULT_BUTTON_CLASSNAME, BUTTON_ANCHOR_CLASSNAME } from '../Button/constants';
 import { AnchorButtonProps, BaseButtonProps, ButtonVariant } from '../Button/types';
 import { useMemo } from 'preact/hooks';
 import { fixedForwardRef, parseBooleanProp, parseClassName } from '../../../utils/preact';
@@ -13,14 +13,16 @@ const AnchorButton = (props: AnchorButtonProps & BaseButtonProps, ref: Ref<HTMLA
     const { click, allChildren, allProps } = useButton(
         classNameValue,
         [...classNameModifiers, variant],
-        DEFAULT_BUTTON_CLASSNAME,
+        `${DEFAULT_BUTTON_CLASSNAME} ${BUTTON_ANCHOR_CLASSNAME}`,
         disabledValue,
         props,
         onClick
     );
 
+    const restProps = useMemo(() => ({ ...allProps, ...(props.onClick && click ? { onClick: click } : {}) }), [click, allProps, props.onClick]);
+
     return (
-        <a {...allProps} href={props.href} ref={ref as Ref<HTMLAnchorElement>} onClick={click}>
+        <a {...restProps} href={props.href} ref={ref as Ref<HTMLAnchorElement>}>
             {allChildren}
         </a>
     );