+++ src/components/internal/FormFields/Select/BalanceAccountSelector/BalanceAccountSelector.test.tsx
@@ -1,10 +1,47 @@
 /**
  * @vitest-environment jsdom
  */
-import { test } from 'vitest';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import { describe, expect, test } from 'vitest';
+import { render, screen, within } from '@testing-library/preact';
+import { BALANCE_ACCOUNTS } from '../../../../../../mocks/mock-data';
 import BalanceAccountSelector from './BalanceAccountSelector';
+import useBalanceAccountSelection from '../../../../../hooks/useBalanceAccountSelection';
 
-test('BalanceAccountSelector', () => {
-    // Write your tests here
+describe('BalanceAccountSelector', () => {
+    type BalanceAccountSelectorWrapperProps = {
+        balanceAccounts: NonNullable<Parameters<typeof useBalanceAccountSelection>[0]>;
+    };
+
+    const BALANCE_ACCOUNTS_WITHOUT_DESCRIPTION = BALANCE_ACCOUNTS.map(({ description, ...account }) => account);
+
+    const BalanceAccountSelectorWrapper = ({ balanceAccounts }: BalanceAccountSelectorWrapperProps) => {
+        const { balanceAccountSelectionOptions } = useBalanceAccountSelection(balanceAccounts);
+        return (
+            <BalanceAccountSelector
+                balanceAccountSelectionOptions={balanceAccountSelectionOptions}
+                activeBalanceAccount={balanceAccounts[0]}
+                onBalanceAccountSelection={() => {}}
+            />
+        );
+    };
+
+    test('should render selected balance account with description (if available)', () => {
+        render(<BalanceAccountSelectorWrapper balanceAccounts={BALANCE_ACCOUNTS} />);
+
+        const activeBalanceAccountDescription = BALANCE_ACCOUNTS[0]?.description!;
+        const buttonTextElement = within(screen.getByRole('button')).getByText(activeBalanceAccountDescription);
+
+        expect(buttonTextElement).toBeInTheDocument();
+        expect(buttonTextElement).toBeVisible();
+    });
+
+    test('should render selected balance account with id (if description is missing)', () => {
+        render(<BalanceAccountSelectorWrapper balanceAccounts={BALANCE_ACCOUNTS_WITHOUT_DESCRIPTION} />);
+
+        const activeBalanceAccountId = BALANCE_ACCOUNTS_WITHOUT_DESCRIPTION[0]?.id!;
+        const buttonTextElement = within(screen.getByRole('button')).getByText(activeBalanceAccountId);
+
+        expect(buttonTextElement).toBeInTheDocument();
+        expect(buttonTextElement).toBeVisible();
+    });
 });