+++ tests/integration/components/transactionsOverview/transactionsOverview.spec.ts
@@ -1,7 +1,8 @@
 import { test as base, expect } from '@playwright/test';
 import { TransactionsOverviewPage } from '../../../models/external-components/transactionsOverview.page';
-import { getTranslatedKey, goToPage } from '../../../utils/utils';
+import { getTranslatedKey, goToStory } from '../../../utils/utils';
 import dotenv from 'dotenv';
+import { CUSTOM_URL_EXAMPLE } from '../../../../stories/utils/constants';
 
 dotenv.config({ path: './envs/.env' });
 
@@ -18,7 +19,7 @@ const test = base.extend<{
 
 test('cells should show correct value and open correct modal ', async ({ transactionsOverviewPage, page }) => {
     const transactionsOverview = transactionsOverviewPage;
-    await goToPage({ page, id: `${COMPONENT_PREFIX}--default` });
+    await goToStory(page, { id: `${COMPONENT_PREFIX}--default` });
     await transactionsOverview.applyDateFilter('2024-01-01');
     await transactionsOverview.getCell('amount').waitFor();
     await transactionsOverview.firstRow.click();
@@ -29,7 +30,7 @@ test('cells should show correct value and open correct modal ', async ({ transac
 
 test.describe('Filters', () => {
     test('all filters should be attached', async ({ transactionsOverviewPage, page }) => {
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--default` });
+        await goToStory(page, { id: `${COMPONENT_PREFIX}--default` });
         const transactionsOverview = transactionsOverviewPage;
         await expect(transactionsOverview.filterBar).toBeAttached();
         await expect(transactionsOverview.dateFilter).toBeAttached();
@@ -38,44 +39,52 @@ test.describe('Filters', () => {
     //TODO test('filtering by date range should work')
 });
 
-test.describe('Transaction List with custom columns', () => {
-    test('Extra columns should be rendered', async ({ transactionsOverviewPage, page }) => {
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--custom-columns` });
-        const transactionsOverview = transactionsOverviewPage;
-        await expect(transactionsOverview.getHeader('Store')).toBeAttached();
-        await expect(transactionsOverview.getHeader('Product')).toBeAttached();
-        await expect(transactionsOverview.getHeader('Reference')).toBeAttached();
-    });
-
-    test('Extra columns values should render with string or the format {value: string}', async ({ transactionsOverviewPage, page }) => {
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--custom-columns` });
+test.describe('Transaction List with data customization', () => {
+    test('should render extra columns', async ({ context, transactionsOverviewPage, page }) => {
+        await goToStory(page, { id: `${COMPONENT_PREFIX}--data-customization` });
         const transactionsOverview = transactionsOverviewPage;
 
-        // _store: { value: string; icon: { url: string } }
+        // _product (TEXT TYPE)
+        const productFirstRow = transactionsOverview.getCell('_product', 0);
+        await expect(productFirstRow).toHaveText('Bubble tea');
+
+        // _store (ICON TYPE)
         const storeFirstRow = transactionsOverview.getCell('_store', 0);
         await expect(storeFirstRow).toHaveText('Sydney');
         await expect(storeFirstRow.getByRole('img')).toBeAttached();
 
-        // _reference: string
+        // _reference: (LINK TYPE)
         const referenceFirstRow = transactionsOverview.getCell('_reference', 0);
-        await expect(referenceFirstRow).toHaveText('8W54BM75W7DYCIVK');
-    });
+        const REFERENCE = '8W54BM75W7DYCIVK';
+        await expect(referenceFirstRow).toHaveText(REFERENCE);
 
-    test('Columns should be reordered', async ({ transactionsOverviewPage, page }) => {
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--custom-order` });
-        const transactionsOverview = transactionsOverviewPage;
-        const headers = transactionsOverview.dataGrid.rootElement.getByRole('columnheader');
-        await expect(headers.nth(0)).toHaveText('Transaction type');
-        await expect(headers.nth(1)).toHaveText('Payment method');
-        await expect(headers.nth(2)).toHaveText('Date');
-        await expect(headers.nth(3)).toHaveText('Amount');
+        //hidden column
+        const transactionTypeFirstRow = transactionsOverview.getCell('transactionType', 0);
+        await expect(transactionTypeFirstRow).not.toBeAttached();
+
+        const [newPage] = await Promise.all([
+            context.waitForEvent('page'), // Waits for a new 'page' event in this browser context
+            referenceFirstRow.click(), // This click opens the link in a new tab
+        ]);
+        await newPage.waitForLoadState();
+        expect(newPage.url()).toContain(CUSTOM_URL_EXAMPLE);
+
+        // _action (BUTTON TYPE)
+        const messages: string[] = [];
+        page.on('console', message => {
+            messages.push(message.text());
+        });
+        const actionFirstRow = transactionsOverview.getCell('_button', 0);
+        await actionFirstRow.getByRole('button').click();
+
+        expect(messages).toContain('Action');
     });
 });
 
 test.describe('Transaction details modal with partial refunds', () => {
     test.beforeEach(async ({ transactionsOverviewPage, page }) => {
         const transactionsOverview = transactionsOverviewPage;
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--default` });
+        await goToStory(page, { id: `${COMPONENT_PREFIX}--default` });
         await transactionsOverview.applyDateFilter('2024-01-01');
         await transactionsOverview.getCell('amount').waitFor();
         await transactionsOverview.firstRow.click();
@@ -101,7 +110,7 @@ test.describe('Transaction details modal with partial refunds', () => {
 test.describe('Refund action modal', () => {
     test.beforeEach(async ({ transactionsOverviewPage, page }) => {
         const transactionsOverview = transactionsOverviewPage;
-        await goToPage({ page, id: `${COMPONENT_PREFIX}--default` });
+        await goToStory(page, { id: `${COMPONENT_PREFIX}--default` });
         await transactionsOverview.applyDateFilter('2024-01-01');
         await transactionsOverview.getCell('amount').waitFor();
         await transactionsOverview.firstRow.click();