+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # @adyen/adyen-platform-experience-web
 
+## 1.4.1
+
+### Fixed
+
+-   Replaced unsupported regex lookbehind patterns causing app to crash for older browsers.
+
+
+## 1.4.0
+
+### New
+
+-   In the Capital Overview component, you can now view bank account details needed to send [unscheduled repayments](https://docs.adyen.com/platforms/capital/capital-components/#capital-overview-component), such as early repayments or to catch up if behind schedule.
+-   You can now add custom data in your [transaction](https://docs.adyen.com/platforms/build-user-dashboards/transaction-components/), [payout](https://docs.adyen.com/platforms/build-user-dashboards/payout-components/), and [report](https://docs.adyen.com/platforms/build-user-dashboards/reports-component/) components, such as your own fields, icons, links, and buttons. Additionally, you have control over the field settings, including their names and visibility.
+
+### Improved
+
+-   In the Capital Offer component, the Request Funds button is now disabled after a successful funds request.
+-   We improved responsiveness of the components by swapping all media queries with container queries.
+
+### Fixed
+
+-   We fixed the issue where the balance account selector in the Payouts, Reports, and Transactions Overview components failed to load if no balance account descriptions were provided.
+
+## 1.3.1
+
+### Patch Changes
+
+-   43352e5: Fix styles import path
+
+## 1.3.0
+
+### Minor Changes
+
+-   865ff93: Add isInUnsupportedRegion state to getState method of Capital component.
+
+### Patch Changes
+
+-   865ff93: Add an interface for the unsupported region which will be shown by default when the user's legal entity region is unsupported for the Capital component.
+
 ## 1.2.0
 
 ### Minor Changes